Modify eval file to add more log statements, correct number of test images, remove unused code
