+++ submodules/KittiObjective2/Makefile
@@ -1,5 +1,5 @@
 make:
-	g++ -O3 -DNDEBUG -o evaluate_object2 evaluate_object.cpp -g -std=c++17 -lstdc++fs
+	g++ -O3 -DNDEBUG -o evaluate_object2 evaluate_object.cpp eval_structs.h -g -std=c++17 -lstdc++fs
 
 install:
 	cp evaluate_object ~/bin
