+++ submodules/KittiObjective2/eval_structs.h
@@ -1,0 +1,56 @@
+#include <math.h>
+#include <strings.h>
+#include <string>
+#include <vector>
+/*=======================================================================
+DATA TYPES FOR EVALUATION
+=======================================================================*/
+using namespace std;
+// holding data needed for precision-recall and precision-aos
+struct tPrData {
+  vector<double> v;           // detection score for computing score thresholds
+  double         similarity;  // orientation similarity
+  int32_t        tp;          // true positives
+  int32_t        fp;          // false positives
+  int32_t        fn;          // false negatives
+  tPrData () :
+    similarity(0), tp(0), fp(0), fn(0) {}
+};
+
+// holding bounding boxes for ground truth and detections
+struct tBox {
+  string  type;     // object type as car, pedestrian or cyclist,...
+  double   x1;      // left corner
+  double   y1;      // top corner
+  double   x2;      // right corner
+  double   y2;      // bottom corner
+  double   alpha;   // image orientation
+  tBox (string type, double x1,double y1,double x2,double y2,double alpha) :
+    type(type),x1(x1),y1(y1),x2(x2),y2(y2),alpha(alpha) {}
+};
+
+// holding ground truth data
+struct tGroundtruth {
+  tBox    box;        // object type, box, orientation
+  double  truncation; // truncation 0..1
+  int32_t occlusion;  // occlusion 0,1,2 (non, partly, fully)
+  tGroundtruth () :
+    box(tBox("invalild",-1,-1,-1,-1,-10)),truncation(-1),occlusion(-1) {}
+  tGroundtruth (tBox box,double truncation,int32_t occlusion) :
+    box(box),truncation(truncation),occlusion(occlusion) {}
+  tGroundtruth (string type,double x1,double y1,double x2,double y2,double alpha,double truncation,int32_t occlusion) :
+    box(tBox(type,x1,y1,x2,y2,alpha)),truncation(truncation),occlusion(occlusion) {}
+};
+
+// holding detection data
+struct tDetection {
+  tBox    box;    // object type, box, orientation
+  double  thresh; // detection score
+  tDetection ():
+    box(tBox("invalid",-1,-1,-1,-1,-10)),thresh(-1000) {}
+  tDetection (tBox box,double thresh) :
+    box(box),thresh(thresh) {}
+  tDetection (string type,double x1,double y1,double x2,double y2,double alpha,double thresh) :
+    box(tBox(type,x1,y1,x2,y2,alpha)),thresh(thresh) {}
+};
+
